feat: add example functions
