+++ README.md
@@ -3,22 +3,23 @@
 Recommend getting the go extension for vscode.
 Go has a built in linter/formatter
 
-Get go with homebrew
+Get go and make if you don't already have them
 
 ```bash
-brew install go
+brew install go make
 ```
 
 Run with
 
 ```bash
-go run main.go
+make
 ```
 
 Build with
 
 ```bash
-go build
+# in the project directory
+make build
 # run binary directly
-./tbd
+./_output/bin/tbd
 ```
