+++ pkg/pull/pull.go
@@ -1,0 +1,7 @@
+package pull
+
+import "fmt"
+
+func Pull() {
+	fmt.Println("pull called")
+}
