+++ pkg/run/run.go
@@ -1,0 +1,7 @@
+package run
+
+import "fmt"
+
+func Run() {
+	fmt.Println("run called")
+}
